// domain/osu/enums/Skill.enum.ts
export enum ESkillType {
    Aim = "Aim",
    Speed = "Speed",
    Accuracy = "Accuracy",
    Stamina = "Stamina",
    Control = "Control",

    Rhythm = "Rhythm",
    Reading = "Reading",
    Colour = "Colour",
    Movement = "Movement",
    Strain = "Strain",
}

export enum ESkillRank {
    Weak = "Weak",
    Common = "Common",
    Average = "Average",
    Expert = "Expert",
    Elite = "Elite",
    Master = "Master",
    Legend = "Legend",
}